requirements.txt
# LLM Providers
anthropic>=0.40.0

# Web & HTTP
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Google APIs
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# Data & Validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Storage
aiosqlite>=0.20.0

# CLI & Terminal
rich>=13.7.0
prompt-toolkit>=3.0.43

# Configuration
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Async
anyio>=4.3.0

# XML parsing (sitemaps)
xmltodict>=0.13.0
